+++ CHANGELOG.md
@@ -5,7 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.0.10] - 2023-10-12
+
+### Fixed
+- Fixed issue where parent records were not getting deleted from solr if empty. (DR-2342)
+- Fixed issue where suppressed captures were not being pulled back from solr. (DR-2470)
+
 ## [1.0.9] - 2023-07-20
+
 ### Updated
 - Added additional rights statements that release high resolution permalinks (DR-2396)